+++ src/Lykke.Service.OperationsRepository.Core/CashOperations/ITransferEventsRepository.cs
@@ -22,7 +22,7 @@ namespace Lykke.Service.OperationsRepository.Core.CashOperations
         public string AddressTo { get; set; }
         public bool? IsSettled { get; set; }
         public TransactionStates State { get; set; }
-        public double FeeSize { get; set; }
+        public decimal FeeSize { get; set; }
         public FeeType FeeType { get; set; }
 
         public static TransferEvent CreateNew(string clientId, string clientMultiSig, string fromId, string assetId, double amount,
